backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiofiles>=23.0.0",
    "asyncpg>=0.30.0",
    "docling[rapidocr]>=2.0.0",
    "fastapi>=0.135.1",
    "google-generativeai>=0.8.0",
    "httpx>=0.27.0",
    "onnxruntime>=1.24.3",
    "pgvector>=0.3.0",
    "pillow>=10.0.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.12",
    "rapidocr>=3.8.1",
    "sentence-transformers>=2.0.0",
    "sqlalchemy>=2.0.0",
    "tiktoken>=0.7.0",
    "uvicorn>=0.42.0",
]
